Remove workaround and allow qualified arns to be used
At one point there was a bug that qualified the lambda invoke arn
unnecessarily but that has been fixed. The workaround is making it so
it's impossible to use a qualified arn for the routes. Being able to
call a specific version of a lambda could be important - especially when
combined with the `deployments` feature of apigw.